Ifpack2 btds gemm (#2829) (#2843)
* Revert "batched/dense/impl/KokkosBatched_Gemm_Team_Impl.hpp (#2626)"

This reverts commit 82605a9b1bd4b34edc5beb63f81f50f9da14fd4b.



* Revert "ConjTrans support for batched team gemm (#2580)"

This reverts commit 386663c2c9a95b28b2838d2deef4134b625ba142.



* Batch - Dense: Re-applying GEMM fixes for extent/stride

We have been liberally querrying the extent and stride of Views
without checking if the rank of the view is high enough to return
a valid value. The fixes lead to failures in BTDS which might point
to a bug that relied on the old UB behavior of Kokkos::View x(



* Applying clang-format



* Batched - Utils: fixing extent and stride calculation

Making sure we return 1 as the minimum stride and extent
even if a view has a rank lower than the stride/extent querried.



* Update batched/KokkosBatched_Util.hpp




* Update batched/KokkosBatched_Util.hpp




---------

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>
Signed-off-by: Luc Berger <lberge@sandia.gov>
Co-authored-by: Luc Berger <lberge@sandia.gov>
Co-authored-by: yasahi-hpc <57478230+yasahi-hpc@users.noreply.github.com>