CI - BDW -> Host: renaming broadwell to host
This makes more sense in my mind as we can still name the jobs
within this file to describe the arch they run on but as things
evolve we won't need to rename files and also the at2.yml will
not need to change as well (less mistakes hopefully).

Also switchin back to CXX=17 for the gcc 10.2.0 job as the compiler
is too low for Core requirements.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>