CI workflows: updating a few workflows to CXX=20 (#2747)
* CI workflows: updating a few worklows to CXX=20

Getting this in our CI ahead of release 5.0 of the
ecosystem which will make it a requirement. The compilers
for github OSX and Linux actions are recent enough (gcc 13 and clang 15)

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* CI - BDW -> Host: renaming broadwell to host

This makes more sense in my mind as we can still name the jobs
within this file to describe the arch they run on but as things
evolve we won't need to rename files and also the at2.yml will
not need to change as well (less mistakes hopefully).

Also switchin back to CXX=17 for the gcc 10.2.0 job as the compiler
is too low for Core requirements.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* CI - Host: more unification of the yaml files

Moving SPR into host and switching build to 4.7.01 even if only
for a short while until 5.0.0 gets released...

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* untabify...

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* CI - Host: removing spr.yml

The job in spr.yml is now included in host.yml so no need to keep
the old file around...

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

---------

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>
(cherry picked from commit fd831208fc8558ee1cab055154f71bc43071ec39)
Signed-off-by: Nathan Ellingwood <ndellin@sandia.gov>