Batched - QR: Implementing a version correct for complex scalars (#2590)
* Batched - QR: Implementing a version correct for complex scalars

tau is computed correctly to obtain a real valued matrix diagonal
for R which is what LAPACK is doing.

Current implementation did not conjugate the scaling factor tau
and did not allow to distinguish between non-transpose and conj
transpose application of a Householder reflection.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Batched - Householder: actually run on device

The low level unit-test was running the functors on host while
debugging. Moving them into a parallel_for to avoid obvious issues
with data access.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Batched - QR: fix Householder unit-tests

The tests were a little sloppy in how it defined data and policies
this should be a more reasonable setup now!

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Apply clang-format

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

* Batched - QR: applied clang-format and fixed docs

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>

---------

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>